# Fields will merge to AST visitor