--- a/Untitled
+++ b/Untitled
@@ -1,0 +1,12 @@
+class Analyzer < Visitor
+  def initialize
+  end
+end
+
+class MyAnalyzer < Analyzer
+  def initial_state
+  end
+
+  def on_field(node, parent, defn)
+  end
+def