--- a/lib/graphql/static_validation/rules/fields_will_merge.rb
+++ b/lib/graphql/static_validation/rules/fields_will_merge.rb
@@ -285,11 +285,11 @@ module GraphQL
             n.arguments.reduce({}) do |memo, a|
               arg_value = a.value
               memo[a.name] = case arg_value
-                             when GraphQL::Language::Nodes::AbstractNode
-                               arg_value.to_query_string
-                             else
-                               GraphQL::Language.serialize(arg_value)
-                             end
+              when GraphQL::Language::Nodes::AbstractNode
+                arg_value.to_query_string
+              else
+                GraphQL::Language.serialize(arg_value)
+              end
               memo
             end
           else