--- a/spec/graphql/static_validation/rules/fields_will_merge_spec.rb
+++ b/spec/graphql/static_validation/rules/fields_will_merge_spec.rb
@@ -373,6 +373,7 @@ describe GraphQL::StaticValidation::FieldsWillMerge do
       }
     |}
 
+    focus
     it "fails rule" do
       expected_errors = [
         {