Converts the FieldsWillMerge validator to a "pure AST" validator, without using the rewrite. Most of the logic is based on the algorithm in graphql-js (https://github.com/graphql/graphql-js/blob/master/src/validation/rules/OverlappingFieldsCanBeMerged.js)

The algorithm is explained through comments in the class. 

  - [x] make the tests pass
  - [x] refactor / beautify logic
  - [x] performance optimizations
  - [x] Fix https://github.com/graphql/graphql-js/pull/1497 but here