@xuorig 2018-08-30
Preliminary benchmarks:

Master:
```
Warming up --------------------------------------
validate - introspection
                        38.000  i/100ms
validate - abstract fragments
                        82.000  i/100ms
validate - abstract fragments 2
                        43.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection
                        394.693  (±11.7%) i/s -      1.938k in   5.004816s
validate - abstract fragments
                        721.422  (±16.5%) i/s -      3.526k in   5.053876s
validate - abstract fragments 2
                        439.557  (± 9.8%) i/s -      2.193k in   5.047115s
validate - big query     39.992  (± 7.5%) i/s -    200.000  in   5.036172s
```

This branch:
```
Warming up --------------------------------------
validate - introspection
                        38.000  i/100ms
validate - abstract fragments
                        82.000  i/100ms
validate - abstract fragments 2
                        43.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection
                        394.693  (±11.7%) i/s -      1.938k in   5.004816s
validate - abstract fragments
                        721.422  (±16.5%) i/s -      3.526k in   5.053876s
validate - abstract fragments 2
                        439.557  (± 9.8%) i/s -      2.193k in   5.047115s
validate - big query     39.992  (± 7.5%) i/s -    200.000  in   5.036172s
```