@rmosolgo 2018-08-30
(For those keeping score at home, about 25% faster 😎 )