@xuorig 2018-08-31
After some caching, hard to say if it's much better, but I think the benches we have don't necessarily require caches so that's probably why.

```
Warming up --------------------------------------
validate - introspection
                        26.000  i/100ms
validate - abstract fragments
                        71.000  i/100ms
validate - abstract fragments 2
                        39.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection
                        384.955  (±13.2%) i/s -      1.898k in   5.054420s
validate - abstract fragments
                        778.436  (±14.1%) i/s -      3.834k in   5.086308s
validate - abstract fragments 2
                        415.047  (±14.2%) i/s -      2.028k in   5.013408s
validate - big query     46.101  (±13.0%) i/s -    228.000  in   5.068996s
```