@rmosolgo 2018-08-31
Yeah, I bet "big query" benefited the most, I think it has lots of fragments.