@xuorig 2018-08-31
And another bench mark, I think this should be ready for review! 

```
Warming up --------------------------------------
validate - introspection
                        40.000  i/100ms
validate - abstract fragments
                        85.000  i/100ms
validate - abstract fragments 2
                        44.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection
                        417.224  (± 8.6%) i/s -      2.080k in   5.030352s
validate - abstract fragments
                        835.254  (± 8.7%) i/s -      4.165k in   5.034208s
validate - abstract fragments 2
                        446.764  (±10.5%) i/s -      2.244k in   5.096434s
validate - big query     43.860  (±16.0%) i/s -    216.000  in   5.086208s
```