@xuorig 2018-09-04
Great context here from graphql-js: https://github.com/graphql/graphql-js/pull/1497#issuecomment-418454489

Sounds like graphql-js doesn't handle these edge cases, and the spec doesn't seem to either 🤔