@xuorig 2018-09-05
After trying a few things, I've just realized that:

```graphql
        {
          pet {
            ... on Dog {
              ... on Pet {
                name
              }
            }
            ... on Cat {
              name(surname: true)
            }
          }
        }
```

Used to be valid (Not sure how internal representation used to handle that case), but is not anymore. The algo probably has to maintain lists of parents instead of only the immediate parent of fields, in this case, `Pet` and `Cat` are not mutually exclusive (conflict), but looking at the grand parent `Dog` we should know that they are in fact exclusive. (no conflict)