@rmosolgo 2018-09-05
The implementation of that was in `InternalRepresentation::Scope`, it kept track of which concrete types were allowed at a given point in the AST