@xuorig 2018-09-05
With new algorithm, minus caching for now. Not a terrible change performance wise it seems

```
Warming up --------------------------------------
validate - introspection
                        41.000  i/100ms
validate - abstract fragments
                        81.000  i/100ms
validate - abstract fragments 2
                        44.000  i/100ms
validate - big query     3.000  i/100ms
Calculating -------------------------------------
validate - introspection
                        408.896  (± 4.4%) i/s -      2.050k in   5.023692s
validate - abstract fragments
                        796.939  (± 7.7%) i/s -      3.969k in   5.014707s
validate - abstract fragments 2
                        448.745  (± 7.6%) i/s -      2.244k in   5.035781s
validate - big query     44.255  (± 4.5%) i/s -    222.000  in   5.022948s
```