@rmosolgo 2018-09-13
Wow, I spent some time reading the code, thanks again for your amazing work here! 🎊  

An interesting take-away for me is how iterative this implementation is. I mean, the comparison really boils down to looping over two sets and comparing each member to each other member. I wouldn't have guessed that it would be so fast, but it sure is. I would have used a Hash, but as shown in any profiler output for GraphQL-Ruby, `Kernel#hash` is a big time-waster of ours! So maybe I should keep an eye out for applying this more often.  Maybe the Ruby VM is faster than the data-structure based approach. 

I have an outstanding question, but it doesn't block merging this PR. Let's say we have a query like this: 

```graphql
{
  a { x y } 
  b { x y } 
}
```

When we enter `a`, we'll deeply check it for conflicts, including checking its subselections (`{x y}`) against `b`'s subselections (`{x y}`). Is that right? 

Then, when we enter `b`, will we also compare _its_ subselections against `a`'s subselections? I see a cache for fragments in a case like this, I wonder if it's worth caching here, too. 

Also, I see that `@fields_and_fragments_from_node` is unused (hopefully not due to my bad git behavior 😩 ) so I will remove it before merging.